app-auth/graphapponlytutorial/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.10
# by the following command:
#
#    pip-compile --output-file=requirements.txt --resolver=backtracking
#
aiohttp==3.8.5
    # via microsoft-kiota-authentication-azure
aiosignal==1.3.1
    # via aiohttp
anyio==3.7.1
    # via httpcore
async-timeout==4.0.2
    # via aiohttp
attrs==23.1.0
    # via aiohttp
azure-core==1.28.0
    # via
    #   azure-identity
    #   microsoft-kiota-authentication-azure
azure-identity==1.13.0
    # via
    #   -r requirements.in
    #   msgraph-sdk
certifi==2023.7.22
    # via
    #   httpcore
    #   httpx
    #   requests
cffi==1.15.1
    # via cryptography
charset-normalizer==3.2.0
    # via
    #   aiohttp
    #   requests
cryptography==41.0.3
    # via
    #   azure-identity
    #   msal
    #   pyjwt
frozenlist==1.4.0
    # via
    #   aiohttp
    #   aiosignal
h11==0.14.0
    # via httpcore
h2==4.1.0
    # via httpx
hpack==4.0.0
    # via h2
httpcore==0.17.3
    # via httpx
httpx[http2]==0.24.1
    # via
    #   microsoft-kiota-http
    #   msgraph-core
hyperframe==6.0.1
    # via h2
idna==3.4
    # via
    #   anyio
    #   httpx
    #   requests
    #   yarl
microsoft-kiota-abstractions==0.6.0
    # via
    #   microsoft-kiota-authentication-azure
    #   microsoft-kiota-http
    #   microsoft-kiota-serialization-json
    #   microsoft-kiota-serialization-text
    #   msgraph-core
    #   msgraph-sdk
microsoft-kiota-authentication-azure==0.2.0
    # via msgraph-sdk
microsoft-kiota-http==0.4.4
    # via msgraph-core
microsoft-kiota-serialization-json==0.3.7
    # via msgraph-sdk
microsoft-kiota-serialization-text==0.2.1
    # via msgraph-sdk
msal==1.23.0
    # via
    #   azure-identity
    #   msal-extensions
msal-extensions==1.0.0
    # via azure-identity
msgraph-core==1.0.0a4
    # via msgraph-sdk
msgraph-sdk==1.0.0a13
    # via -r requirements.in
multidict==6.0.4
    # via
    #   aiohttp
    #   yarl
portalocker==2.7.0
    # via msal-extensions
pycparser==2.21
    # via cffi
pyjwt[crypto]==2.8.0
    # via msal
python-dateutil==2.8.2
    # via
    #   microsoft-kiota-serialization-json
    #   microsoft-kiota-serialization-text
requests==2.31.0
    # via
    #   azure-core
    #   msal
six==1.16.0
    # via
    #   azure-core
    #   azure-identity
    #   python-dateutil
sniffio==1.3.0
    # via
    #   anyio
    #   httpcore
    #   httpx
typing-extensions==4.7.1
    # via azure-core
uritemplate==4.1.1
    # via microsoft-kiota-abstractions
urllib3==2.0.4
    # via requests
yarl==1.9.2
    # via aiohttp
